# Improve performance of `Jisx0402::Tree::Node#values`